Crate halo2_base

Base library to build Halo2 circuits.

Re-exports

pub use halo2_proofs_axiom as halo2_proofs;

Modules

gates

Module that contains the main API for creating and working with circuits.

safe_types

Module for SafeType which enforce value range and realted functions.

utils

Utility functions for converting between different types of field elements.

Structs

AssignedValue

Pointer containing cell value and location within Context.

Context

Represents a single thread of an execution trace.

ContextCell

Pointer to the position of a cell at offset in an advice column within a Context of context_id.

Enums

QuantumCell

Convenience Enum which abstracts the scenarios under a value is added to an advice column.

Constants

SKIP_FIRST_PASS

Constant representing whether the Layouter calls synthesize once just to get region shape.