Module halo2_base::gates

Module that contains the main API for creating and working with circuits.

Re-exports

pub use flex_gate::GateChip;

pub use flex_gate::GateInstructions;

pub use range::RangeChip;

pub use range::RangeInstructions;

Modules

builder

Module that helps auto-build circuits

flex_gate

Module implementing our simple custom gate and common functions using it

range

Module using a single lookup table for range checks