Function halo2_base::gates::builder::assign_threads_in

pub fn assign_threads_in<F: ScalarField>(
    phase: usize,
    threads: Vec<Context<F>>,
    config: &FlexGateConfig<F>,
    lookup_advice: &[Column<Advice>],
    region: &mut Region<'_, F>,
    break_points: ThreadBreakPoints
)

Assigns threads to regions of advice column.

Uses preprocessed break_points to assign where to divide the advice column into a new column for each thread.

Performs only witness generation, so should only be evoked during proving not keygen.

Assumes that the advice columns are already assigned.

phase - the phase of the circuit

threads - Vec threads to assign

config - immutable reference to the configuration of the circuit

lookup_advice - Slice of lookup advice columns

region - mutable reference to the region to assign threads to

break_points - the preprocessed break points for the threads