Struct halo2_base::gates::builder::GateThreadBuilder

pub struct GateThreadBuilder<F: ScalarField> {
    pub threads: [Vec<Context<F>>; 3],
    pub witness_gen_only: bool,
    /* private fields */
}

Builds the process for gate threading

Fields
threads: [Vec<Context<F>>; 3]

Threads for each challenge phase

witness_gen_only: bool

Flag for witness generation. If true, the gate thread builder is used for witness generation only.

Implementations

impl<F: ScalarField> GateThreadBuilder<F>

pub fn new(witness_gen_only: bool) -> Self

Creates a new GateThreadBuilder and spawns a main thread in phase 0.

witness_gen_only: If true, the GateThreadBuilder is used for witness generation only.

If true, the gate thread builder only does witness asignments and does not store constraint information – this should only be used for the real prover.

If false, the gate thread builder is used for keygen and mock prover (it can also be used for real prover) and the builder stores circuit information (e.g. copy constraints, fixed columns, enabled selectors).

These values are fixed for the circuit at key generation time, and they do not need to be re-computed by the prover in the actual proving phase.

pub fn mock() -> Self

Creates a new GateThreadBuilder with witness_gen_only set to false.

Performs the witness assignment computations and then checks using normal programming logic whether the gate constraints are all satisfied.

pub fn keygen() -> Self

Creates a new GateThreadBuilder with witness_gen_only set to false.

Performs the witness assignment computations and generates prover and verifier keys.

pub fn prover() -> Self

Creates a new GateThreadBuilder with witness_gen_only set to true.

Performs the witness assignment computations and then runs the proving system.

pub fn unknown(self, use_unknown: bool) -> Self

Creates a new GateThreadBuilder with use_unknown flag set.

use_unknown: If true, during key generation witness [Value]s are replaced with Value::unknown() for safety.

pub fn main(&mut self, phase: usize) -> &mut Context<F>

Returns a mutable reference to the Context of a gate thread. Spawns a new thread for the given phase, if none exists.

phase: The challenge phase (as an index) of the gate thread.

pub fn witness_gen_only(&self) -> bool

Returns the witness_gen_only flag.

pub fn use_unknown(&self) -> bool

Returns the use_unknown flag.

pub fn thread_count(&self) -> usize

Returns the current number of threads in the GateThreadBuilder.

pub fn get_new_thread_id(&mut self) -> usize

Creates a new thread id by incrementing the thread count

pub fn new_thread(&mut self, phase: usize) -> &mut Context<F>

Spawns a new thread for a new given phase. Returns a mutable reference to the Context of the new thread.

phase: The phase (index) of the gate thread.

pub fn config(
&self,
k: usize,
minimum_rows: Option<usize>
) -> FlexGateConfigParams

Auto-calculates configuration parameters for the circuit

k: The number of in the circuit (i.e. numeber of rows = 2k)

minimum_rows: The minimum number of rows in the circuit that cannot be used for witness assignments and contain random blinding factors to ensure zk property, defaults to 0.

pub fn assign_all(
&self,
config: &FlexGateConfig<F>,
lookup_advice: &[Vec<Column<Advice>>],
q_lookup: &[Option<Selector>],
region: &mut Region<'_, F>,
_: KeygenAssignments<F>
) -> KeygenAssignments<F>

Assigns all advice and fixed cells, turns on selectors, and imposes equality constraints.

Returns the assigned advices, and constants in the form of KeygenAssignments.

Assumes selector and advice columns are already allocated and of the same length.

Note: assign_all() should be called during keygen or if using mock prover. It also works for the real prover, but there it is more optimal to use assign_threads_in instead.

config: The FlexGateConfig of the circuit.

lookup_advice: The lookup advice columns.

q_lookup: The lookup advice selectors.

region: The [Region] of the circuit.

assigned_advices: The assigned advice cells.

assigned_constants: The assigned fixed cells.

break_points: The break points of the circuit.