Module halo2_base::gates::flex_gate

Module implementing our simple custom gate and common functions using it

Structs

BasicGateConfig

A configuration for a basic gate chip describing the selector, and advice column values.

FlexGateConfig

Defines a configuration for a flex gate chip describing the selector, and advice column values for the chip.

GateChip

A chip that implements the GateInstructions trait supporting basic arithmetic operations.

Enums

GateStrategy

Specifies the gate strategy for the gate chip

Constants

MAX_PHASE

The maximum number of phases in halo2.

Traits

GateInstructions

Trait that defines basic arithmetic operations for a gate.