Trait halo2_base::gates::flex_gate::GateInstructions

pub trait GateInstructions<F: ScalarField> {

    fn strategy(&self) -> GateStrategy;
    fn pow_of_two(&self) -> &[F];
    fn get_field_element(&self, n: u64) -> F;
    fn inner_product<QA>(
        &self,
        ctx: &mut Context<F>,
        a: impl IntoIterator<Item = QA>,
        b: impl IntoIterator<Item = QuantumCell<F>>
    ) -> AssignedValue<F>
    where
        QA: Into<QuantumCell<F>>;
    fn inner_product_left_last<QA>(
        &self,
        ctx: &mut Context<F>,
        a: impl IntoIterator<Item = QA>,
        b: impl IntoIterator<Item = QuantumCell<F>>
    ) -> (AssignedValue<F>, AssignedValue<F>)
    where
        QA: Into<QuantumCell<F>>;
    fn inner_product_with_sums<'thread, QA>(
        &self,
        ctx: &'thread mut Context<F>,
        a: impl IntoIterator<Item = QA>,
        b: impl IntoIterator<Item = QuantumCell<F>>
    ) -> Box<dyn Iterator<Item = AssignedValue<F>> + 'thread>
    where
        QA: Into<QuantumCell<F>>;
    fn sum_products_with_coeff_and_var(
        &self,
        ctx: &mut Context<F>,
        values: impl IntoIterator<Item = (F, QuantumCell<F>, QuantumCell<F>)>,
        var: QuantumCell<F>
    ) -> AssignedValue<F>;
    fn select(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>,
        sel: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F>;
    fn or_and(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>,
        c: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F>;
    fn num_to_bits(
        &self,
        ctx: &mut Context<F>,
        a: AssignedValue<F>,
        range_bits: usize
    ) -> Vec<AssignedValue<F>>;

    fn add(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn sub(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn neg(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn mul(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn mul_add(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>,
        c: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn mul_not(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn assert_bit(&self, ctx: &mut Context<F>, x: AssignedValue<F>) { ... }
    fn div_unsafe(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn assert_is_const(
        &self,
        ctx: &mut Context<F>,
        a: &AssignedValue<F>,
        constant: &F
    ) { ... }
    fn sum<Q>(
        &self,
        ctx: &mut Context<F>,
        a: impl IntoIterator<Item = Q>
    ) -> AssignedValue<F>
    where
        Q: Into<QuantumCell<F>>,
    { ... }
    fn partial_sums<'thread, Q>(
        &self,
        ctx: &'thread mut Context<F>,
        a: impl IntoIterator<Item = Q>
    ) -> Box<dyn Iterator<Item = AssignedValue<F>> + 'thread>
    where
        Q: Into<QuantumCell<F>>,
    { ... }
    fn accumulated_product<QA, QB>(
        &self,
        ctx: &mut Context<F>,
        a: impl IntoIterator<Item = QA>,
        b: impl IntoIterator<Item = QB>
    ) -> Vec<AssignedValue<F>>
    where
        QA: Into<QuantumCell<F>>,
        QB: Into<QuantumCell<F>>,
    { ... }
    fn or(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn and(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn not(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn bits_to_indicator(
        &self,
        ctx: &mut Context<F>,
        bits: &[AssignedValue<F>]
    ) -> Vec<AssignedValue<F>> { ... }
    fn idx_to_indicator(
        &self,
        ctx: &mut Context<F>,
        idx: impl Into<QuantumCell<F>>,
        len: usize
    ) -> Vec<AssignedValue<F>> { ... }
    fn select_by_indicator<Q>(
        &self,
        ctx: &mut Context<F>,
        a: impl IntoIterator<Item = Q>,
        indicator: impl IntoIterator<Item = AssignedValue<F>>
    ) -> AssignedValue<F>
    where
        Q: Into<QuantumCell<F>>,
    { ... }
    fn select_from_idx<Q>(
        &self,
        ctx: &mut Context<F>,
        cells: impl IntoIterator<Item = Q>,
        idx: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F>
    where
        Q: Into<QuantumCell<F>>,
    { ... }
    fn is_zero(
        &self,
        ctx: &mut Context<F>,
        a: AssignedValue<F>
    ) -> AssignedValue<F> { ... }
    fn is_equal(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>
    ) -> AssignedValue<F> { ... }
    fn lagrange_and_eval(
        &self,
        ctx: &mut Context<F>,
        coords: &[(AssignedValue<F>, AssignedValue<F>)],
        x: AssignedValue<F>
    ) -> (AssignedValue<F>, AssignedValue<F>) { ... }

}

Required Methods

fn strategy(&self) -> GateStrategy

fn pow_of_two(&self) -> &[F]

fn get_field_element(&self, n: u64) -> F

fn inner_product<QA>(
&self,
ctx: &mut Context<F>,
a: impl IntoIterator<Item = QA>,
b: impl IntoIterator<Item = QuantumCell<F>>
) -> AssignedValue<F>where
QA: Into<QuantumCell<F>>,

Returns the inner product of <a, b>

fn inner_product_left_last<QA>(
&self,
ctx: &mut Context<F>,
a: impl IntoIterator<Item = QA>,
b: impl IntoIterator<Item = QuantumCell<F>>
) -> (AssignedValue<F>, AssignedValue<F>)where
QA: Into<QuantumCell<F>>,

Returns the inner product of <a, b> and the last item of a after it is assigned

fn inner_product_with_sums<'thread, QA>(
&self,
ctx: &'thread mut Context<F>,
a: impl IntoIterator<Item = QA>,
b: impl IntoIterator<Item = QuantumCell<F>>
) -> Box<dyn Iterator<Item = AssignedValue<F>> + 'thread>where
QA: Into<QuantumCell<F>>,

Returns a vector with the partial sums sum_{j=0..=i} a[j] * b[j].

fn sum_products_with_coeff_and_var(
&self,
ctx: &mut Context<F>,
values: impl IntoIterator<Item = (F, QuantumCell<F>, QuantumCell<F>)>,
var: QuantumCell<F>
) -> AssignedValue<F>

fn select(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>,
sel: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

assumes sel is boolean returns a * sel + b * (1 - sel)

fn or_and(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>,
c: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

returns: a || (b && c)

fn num_to_bits(
&self,
ctx: &mut Context<F>,
a: AssignedValue<F>,
range_bits: usize
) -> Vec<AssignedValue<F>>

returns little-endian bit vectors

Provided Methods

fn add(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

Copies a, b and constrains a + b * 1 = out

fn sub(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

Copies a, b and constrains a + b * (-1) = out

fn neg(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

fn mul(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

Copies a, b and constrains 0 + a * b = out

fn mul_add(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>,
c: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

a * b + c

fn mul_not(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

(1 - a) * b = b - a * b

fn assert_bit(&self, ctx: &mut Context<F>, x: AssignedValue<F>)

Constrain x is 0 or 1.

fn div_unsafe(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

fn assert_is_const(&self, ctx: &mut Context<F>, a: &AssignedValue<F>, constant: &F)

fn sum<Q>(
&self,
ctx: &mut Context<F>,
a: impl IntoIterator<Item = Q>
) -> AssignedValue<F>where
Q: Into<QuantumCell<F>>,

fn partial_sums<'thread, Q>(
&self,
ctx: &'thread mut Context<F>,
a: impl IntoIterator<Item = Q>
) -> Box<dyn Iterator<Item = AssignedValue<F>> + 'thread>where
Q: Into<QuantumCell<F>>,

Returns the assignment trace where output[i] has the running sum sum_{j=0..=i} a[j]

fn accumulated_product<QA, QB>(
&self,
ctx: &mut Context<F>,
a: impl IntoIterator<Item = QA>,
b: impl IntoIterator<Item = QB>
) -> Vec<AssignedValue<F>>where
QA: Into<QuantumCell<F>>,
QB: Into<QuantumCell<F>>,

fn or(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

fn and(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

fn not(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

fn bits_to_indicator(
&self,
ctx: &mut Context<F>,
bits: &[AssignedValue<F>]
) -> Vec<AssignedValue<F>>

assume bits has boolean values returns vec[idx] with vec[idx] = 1 if and only if bits == idx as a binary number

fn idx_to_indicator(
&self,
ctx: &mut Context<F>,
idx: impl Into<QuantumCell<F>>,
len: usize
) -> Vec<AssignedValue<F>>

fn select_by_indicator<Q>(
&self,
ctx: &mut Context<F>,
a: impl IntoIterator<Item = Q>,
indicator: impl IntoIterator<Item = AssignedValue<F>>
) -> AssignedValue<F>where
Q: Into<QuantumCell<F>>,

Assumes for witness generation that only one element of indicator has non-zero value and that value is F::one().

fn select_from_idx<Q>(
&self,
ctx: &mut Context<F>,
cells: impl IntoIterator<Item = Q>,
idx: impl Into<QuantumCell<F>>
) -> AssignedValue<F>where
Q: Into<QuantumCell<F>>,

fn is_zero(&self, ctx: &mut Context<F>, a: AssignedValue<F>) -> AssignedValue<F>

fn is_equal(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>
) -> AssignedValue<F>

fn lagrange_and_eval(
&self,
ctx: &mut Context<F>,
coords: &[(AssignedValue<F>, AssignedValue<F>)],
x: AssignedValue<F>
) -> (AssignedValue<F>, AssignedValue<F>)

given pairs coords[i] = (x_i, y_i), let f be the unique degree len(coords) polynomial such that f(x_i) = y_i for all i.

input: coords, x

output: (f(x), Prod_i (x - x_i))

constrains all x_i and x are distinct