Trait halo2_base::gates::GateInstructions

pub trait GateInstructions<F: ScalarField> {

    fn strategy(&self) -> GateStrategy;
    fn context_id(&self) -> usize;
    fn pow_of_two(&self) -> &[F];
    fn get_field_element(&self, n: u64) -> F;
    fn assign_region_in<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
        phase: usize
    ) -> Vec<AssignedValue<'b, F>>;
    fn assign_region_last_in<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
        phase: usize
    ) -> AssignedValue<'b, F>;
    fn inner_product<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        b: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>
    ) -> AssignedValue<'b, F>;
    fn inner_product_left<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        b: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        a_assigned: &mut Vec<AssignedValue<'b, F>>
    ) -> AssignedValue<'b, F>;
    fn inner_product_with_sums<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        b: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>
    ) -> Box<dyn Iterator<Item = AssignedValue<'b, F>> + 'b>;
    fn sum_products_with_coeff_and_var<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        values: impl IntoIterator<Item = (F, QuantumCell<'a, 'b, F>, QuantumCell<'a, 'b, F>)>,
        var: QuantumCell<'a, 'b, F>
    ) -> AssignedValue<'b, F>;
    fn select<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>,
        sel: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F>;
    fn or_and<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>,
        c: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F>;
    fn num_to_bits<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: &AssignedValue<'v, F>,
        range_bits: usize
    ) -> Vec<AssignedValue<'v, F>>;

    fn assign_region<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>
    ) -> Vec<AssignedValue<'b, F>> { ... }
    fn assign_region_last<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>
    ) -> AssignedValue<'b, F> { ... }
    fn assign_region_smart<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        gate_offsets: impl IntoIterator<Item = usize>,
        equality_offsets: impl IntoIterator<Item = (usize, usize)>,
        external_equality: Vec<(&AssignedValue<'_, F>, usize)>
    ) -> Vec<AssignedValue<'b, F>> { ... }
    fn assign_witnesses<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        witnesses: impl IntoIterator<Item = Value<F>>
    ) -> Vec<AssignedValue<'v, F>> { ... }
    fn load_witness<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        witness: Value<F>
    ) -> AssignedValue<'v, F> { ... }
    fn load_constant<'a>(
        &self,
        ctx: &mut Context<'_, F>,
        c: F
    ) -> AssignedValue<'a, F> { ... }
    fn load_zero<'a>(&self, ctx: &mut Context<'a, F>) -> AssignedValue<'a, F> { ... }
    fn add<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn sub<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn neg<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn mul<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn mul_add<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>,
        c: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn mul_not<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn assert_bit(&self, ctx: &mut Context<'_, F>, x: &AssignedValue<'_, F>) { ... }
    fn div_unsafe<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn assert_equal(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, '_, F>,
        b: QuantumCell<'_, '_, F>
    ) { ... }
    fn assert_is_const(
        &self,
        ctx: &mut Context<'_, F>,
        a: &AssignedValue<'_, F>,
        constant: F
    ) { ... }
    fn sum<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'b, F>,
        a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>
    ) -> AssignedValue<'b, F> { ... }
    fn sum_with_assignments<'a, 'b: 'a>(
        &self,
        ctx: &mut Context<'b, F>,
        a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>
    ) -> Vec<AssignedValue<'b, F>> { ... }
    fn accumulated_product<'a, 'v: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        a: impl IntoIterator<Item = QuantumCell<'a, 'v, F>>,
        b: impl IntoIterator<Item = QuantumCell<'a, 'v, F>>
    ) -> Vec<AssignedValue<'v, F>> { ... }
    fn or<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn and<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn not<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn bits_to_indicator<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        bits: &[AssignedValue<'v, F>]
    ) -> Vec<AssignedValue<'v, F>> { ... }
    fn idx_to_indicator<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        idx: QuantumCell<'_, 'v, F>,
        len: usize
    ) -> Vec<AssignedValue<'v, F>> { ... }
    fn select_by_indicator<'a, 'i, 'b: 'a + 'i>(
        &self,
        ctx: &mut Context<'_, F>,
        a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
        indicator: impl IntoIterator<Item = &'i AssignedValue<'b, F>>
    ) -> AssignedValue<'b, F> { ... }
    fn select_from_idx<'a, 'v: 'a>(
        &self,
        ctx: &mut Context<'_, F>,
        cells: impl IntoIterator<Item = QuantumCell<'a, 'v, F>>,
        idx: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn is_zero<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: &AssignedValue<'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn is_equal<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        a: QuantumCell<'_, 'v, F>,
        b: QuantumCell<'_, 'v, F>
    ) -> AssignedValue<'v, F> { ... }
    fn lagrange_and_eval<'v>(
        &self,
        ctx: &mut Context<'_, F>,
        coords: &[(AssignedValue<'v, F>, AssignedValue<'v, F>)],
        x: &AssignedValue<'v, F>
    ) -> (AssignedValue<'v, F>, AssignedValue<'v, F>) { ... }

}

Required Methods

fn strategy(&self) -> GateStrategy

fn context_id(&self) -> usize

fn pow_of_two(&self) -> &[F]

fn get_field_element(&self, n: u64) -> F

fn assign_region_in<'a, 'b: 'a>(
&self,
ctx: &mut Context<'_, F>,
inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
phase: usize
) -> Vec<AssignedValue<'b, F>>

fn assign_region_last_in<'a, 'b: 'a>(
&self,
ctx: &mut Context<'_, F>,
inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
phase: usize
) -> AssignedValue<'b, F>

fn inner_product<'a, 'b: 'a>(
&self,
ctx: &mut Context<'_, F>,
a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
b: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>
) -> AssignedValue<'b, F>

Returns (assignments, output) where output is the inner product of <a, b>

assignments is for internal use

fn inner_product_left<'a, 'b: 'a>(
&self,
ctx: &mut Context<'_, F>,
a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
b: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
a_assigned: &mut Vec<AssignedValue<'b, F>>
) -> AssignedValue<'b, F>

very specialized for optimal range check, not for general consumption

a_assigned is expected to have capacity a.len()

we re-use a_assigned to save memory allocation

fn inner_product_with_sums<'a, 'b: 'a>(
&self,
ctx: &mut Context<'_, F>,
a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
b: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>
) -> Box<dyn Iterator<Item = AssignedValue<'b, F>> + 'b>

Returns an iterator with the partial sums sum_{j=0..=i} a[j] * b[j].

fn sum_products_with_coeff_and_var<'a, 'b: 'a>(
&self,
ctx: &mut Context<'_, F>,
values: impl IntoIterator<Item = (F, QuantumCell<'a, 'b, F>, QuantumCell<'a, 'b, F>)>,
var: QuantumCell<'a, 'b, F>
) -> AssignedValue<'b, F>

fn select<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>,
sel: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

fn or_and<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>,
c: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

fn num_to_bits<'v>(
&self,
ctx: &mut Context<'_, F>,
a: &AssignedValue<'v, F>,
range_bits: usize
) -> Vec<AssignedValue<'v, F>>

Provided Methods

fn assign_region<'a, 'b: 'a>(
&self,
ctx: &mut Context<'_, F>,
inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>
) -> Vec<AssignedValue<'b, F>>

fn assign_region_last<'a, 'b: 'a>(
&self,
ctx: &mut Context<'_, F>,
inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>
) -> AssignedValue<'b, F>

Only returns the last assigned cell

Does not collect the vec, saving heap allocation

fn assign_region_smart<'a, 'b: 'a>(
&self,
ctx: &mut Context<'_, F>,
inputs: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
gate_offsets: impl IntoIterator<Item = usize>,
equality_offsets: impl IntoIterator<Item = (usize, usize)>,
external_equality: Vec<(&AssignedValue<'_, F>, usize)>
) -> Vec<AssignedValue<'b, F>>

Only call this if ctx.region is not in shape mode, i.e., if not using simple layouter or ctx.first_pass = false

All indices in gate_offsets, equality_offsets, external_equality are with respect to inputs indices

gate_offsets specifies indices to enable selector for the gate; assume gate_offsets is sorted in increasing order

equality_offsets specifies pairs of indices to constrain equality

external_equality specifies an existing cell to constrain equality with the cell at a certain index

fn assign_witnesses<'v>(
&self,
ctx: &mut Context<'_, F>,
witnesses: impl IntoIterator<Item = Value<F>>
) -> Vec<AssignedValue<'v, F>>

fn load_witness<'v>(
&self,
ctx: &mut Context<'_, F>,
witness: Value<F>
) -> AssignedValue<'v, F>

fn load_constant<'a>(
&self,
ctx: &mut Context<'_, F>,
c: F
) -> AssignedValue<'a, F>

fn load_zero<'a>(&self, ctx: &mut Context<'a, F>) -> AssignedValue<'a, F>

fn add<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

Copies a, b and constrains a + b * 1 = out

fn sub<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

Copies a, b and constrains a + b * (-1) = out

fn neg<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

fn mul<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

Copies a, b and constrains 0 + a * b = out

fn mul_add<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>,
c: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

a * b + c

fn mul_not<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

(1 - a) * b = b - a * b

fn assert_bit(&self, ctx: &mut Context<'_, F>, x: &AssignedValue<'_, F>)

Constrain x is 0 or 1.

fn div_unsafe<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

fn assert_equal(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, '_, F>,
b: QuantumCell<'_, '_, F>
)

fn assert_is_const(
&self,
ctx: &mut Context<'_, F>,
a: &AssignedValue<'_, F>,
constant: F
)

fn sum<'a, 'b: 'a>(
&self,
ctx: &mut Context<'b, F>,
a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>
) -> AssignedValue<'b, F>

fn sum_with_assignments<'a, 'b: 'a>(
&self,
ctx: &mut Context<'b, F>,
a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>
) -> Vec<AssignedValue<'b, F>>

Returns the assignment trace where output[3 * i] has the running sum sum_{j=0..=i} a[j]

fn accumulated_product<'a, 'v: 'a>(
&self,
ctx: &mut Context<'_, F>,
a: impl IntoIterator<Item = QuantumCell<'a, 'v, F>>,
b: impl IntoIterator<Item = QuantumCell<'a, 'v, F>>
) -> Vec<AssignedValue<'v, F>>

fn or<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

fn and<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

fn not<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

fn bits_to_indicator<'v>(
&self,
ctx: &mut Context<'_, F>,
bits: &[AssignedValue<'v, F>]
) -> Vec<AssignedValue<'v, F>>

assume bits has boolean values returns vec[idx] with vec[idx] = 1 if and only if bits == idx as a binary number

fn idx_to_indicator<'v>(
&self,
ctx: &mut Context<'_, F>,
idx: QuantumCell<'_, 'v, F>,
len: usize
) -> Vec<AssignedValue<'v, F>>

fn select_by_indicator<'a, 'i, 'b: 'a + 'i>(
&self,
ctx: &mut Context<'_, F>,
a: impl IntoIterator<Item = QuantumCell<'a, 'b, F>>,
indicator: impl IntoIterator<Item = &'i AssignedValue<'b, F>>
) -> AssignedValue<'b, F>

Assumes for witness generation that only one element of indicator has non-zero value and that value is F::one().

fn select_from_idx<'a, 'v: 'a>(
&self,
ctx: &mut Context<'_, F>,
cells: impl IntoIterator<Item = QuantumCell<'a, 'v, F>>,
idx: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

fn is_zero<'v>(
&self,
ctx: &mut Context<'_, F>,
a: &AssignedValue<'v, F>
) -> AssignedValue<'v, F>

fn is_equal<'v>(
&self,
ctx: &mut Context<'_, F>,
a: QuantumCell<'_, 'v, F>,
b: QuantumCell<'_, 'v, F>
) -> AssignedValue<'v, F>

fn lagrange_and_eval<'v>(
&self,
ctx: &mut Context<'_, F>,
coords: &[(AssignedValue<'v, F>, AssignedValue<'v, F>)],
x: &AssignedValue<'v, F>
) -> (AssignedValue<'v, F>, AssignedValue<'v, F>)

given pairs coords[i] = (x_i, y_i), let f be the unique degree len(coords) polynomial such that f(x_i) = y_i for all i.

input: coords, x

output: (f(x), Prod_i (x - x_i))

constrains all x_i and x are distinct