Trait halo2_base::gates::range::RangeInstructions

pub trait RangeInstructions<F: ScalarField> {
    type Gate: GateInstructions<F>;

    fn gate(&self) -> &Self::Gate;
    fn strategy(&self) -> RangeStrategy;
    fn lookup_bits(&self) -> usize;
    fn range_check(
        &self,
        ctx: &mut Context<F>,
        a: AssignedValue<F>,
        range_bits: usize
    );
    fn check_less_than(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>,
        num_bits: usize
    );
    fn is_less_than(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>,
        num_bits: usize
    ) -> AssignedValue<F>;

    fn check_less_than_safe(
        &self,
        ctx: &mut Context<F>,
        a: AssignedValue<F>,
        b: u64
    ) { ... }
    fn check_big_less_than_safe(
        &self,
        ctx: &mut Context<F>,
        a: AssignedValue<F>,
        b: BigUint
    )
    where
        F: BigPrimeField,
    { ... }
    fn is_less_than_safe(
        &self,
        ctx: &mut Context<F>,
        a: AssignedValue<F>,
        b: u64
    ) -> AssignedValue<F> { ... }
    fn is_big_less_than_safe(
        &self,
        ctx: &mut Context<F>,
        a: AssignedValue<F>,
        b: BigUint
    ) -> AssignedValue<F>
    where
        F: BigPrimeField,
    { ... }
    fn div_mod(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<BigUint>,
        a_num_bits: usize
    ) -> (AssignedValue<F>, AssignedValue<F>)
    where
        F: BigPrimeField,
    { ... }
    fn div_mod_var(
        &self,
        ctx: &mut Context<F>,
        a: impl Into<QuantumCell<F>>,
        b: impl Into<QuantumCell<F>>,
        a_num_bits: usize,
        b_num_bits: usize
    ) -> (AssignedValue<F>, AssignedValue<F>)
    where
        F: BigPrimeField,
    { ... }
    fn get_last_bit(
        &self,
        ctx: &mut Context<F>,
        a: AssignedValue<F>,
        limb_bits: usize
    ) -> AssignedValue<F> { ... }

}

Required Associated Types

type Gate: GateInstructions<F>

Required Methods

fn gate(&self) -> &Self::Gate

fn strategy(&self) -> RangeStrategy

fn lookup_bits(&self) -> usize

fn range_check(&self, ctx: &mut Context<F>, a: AssignedValue<F>, range_bits: usize)

Constrain that a lies in the range [0, 2range_bits).

fn check_less_than(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>,
num_bits: usize
)

fn is_less_than(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>,
num_bits: usize
) -> AssignedValue<F>

Returns whether a is in [0, b).

Warning: This may fail silently if a or b have more than num_bits bits

Provided Methods

fn check_less_than_safe(&self, ctx: &mut Context<F>, a: AssignedValue<F>, b: u64)

Checks that a is in [0, b).

Does not require bit assumptions on a, b because we range check that a has at most bit_length(b) bits.

fn check_big_less_than_safe(
&self,
ctx: &mut Context<F>,
a: AssignedValue<F>,
b: BigUint
)where
F: BigPrimeField,

Checks that a is in [0, b).

Does not require bit assumptions on a, b because we range check that a has at most bit_length(b) bits.

fn is_less_than_safe(
&self,
ctx: &mut Context<F>,
a: AssignedValue<F>,
b: u64
) -> AssignedValue<F>

Returns whether a is in [0, b).

Does not require bit assumptions on a, b because we range check that a has at most range_bits bits.

fn is_big_less_than_safe(
&self,
ctx: &mut Context<F>,
a: AssignedValue<F>,
b: BigUint
) -> AssignedValue<F>where
F: BigPrimeField,

Returns whether a is in [0, b).

Does not require bit assumptions on a, b because we range check that a has at most range_bits bits.

fn div_mod(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<BigUint>,
a_num_bits: usize
) -> (AssignedValue<F>, AssignedValue<F>)where
F: BigPrimeField,

Returns (c, r) such that a = b * c + r.

Assumes that b != 0.

fn div_mod_var(
&self,
ctx: &mut Context<F>,
a: impl Into<QuantumCell<F>>,
b: impl Into<QuantumCell<F>>,
a_num_bits: usize,
b_num_bits: usize
) -> (AssignedValue<F>, AssignedValue<F>)where
F: BigPrimeField,

Returns (c, r) such that a = b * c + r.

Assumes that b != 0.

Let X = 2 ** b_num_bits. Write a = a1 * X + a0 and c = c1 * X + c0. If we write b * c0 + r = d1 * X + d0 then b * c + r = (b * c1 + d1) * X + d0

fn get_last_bit(
&self,
ctx: &mut Context<F>,
a: AssignedValue<F>,
limb_bits: usize
) -> AssignedValue<F>

Assume a has been range checked already to limb_bits bits